Move the label-derivation helpers into helpers.ts
getLabelFromChildren and getNodeText lived in Dropdown.tsx; they belong with
the other item/label helpers in helpers.ts. Converted to the file's arrow-const
style, with getNodeText staying module-local (only getLabelFromChildren is
consumed outside).
